Add assertions to multiple classes
Add an assertion in Luna method to ensure input is not empty.
Add an assertion in Main method to ensure stage is initialized before use.
Add an assertion in MainWindow method to ensure response is not empty.
Add an assertion in Storage constructor to ensure filePath parameter is not empty.

These are several functions that will not work properly without proper input and arguments.
These cases are should not happen at all and thus should cause assertion failures.